bridge_circuit_host
Module bridge_circuit_host

Constants

MAINNET_BRIDGE_CIRCUIT_ELF

MAINNET_HEADER_CHAIN_ELF

MAINNET_WORK_ONLY_ELF

REGTEST_BRIDGE_CIRCUIT_ELF

REGTEST_HEADER_CHAIN_ELF

REGTEST_WORK_ONLY_ELF

SIGNET_BRIDGE_CIRCUIT_ELF

SIGNET_HEADER_CHAIN_ELF

SIGNET_WORK_ONLY_ELF

TESTNET4_BRIDGE_CIRCUIT_ELF

TESTNET4_HEADER_CHAIN_ELF

TESTNET4_HEADER_CHAIN_GUEST_ELF

TESTNET4_WORK_ONLY_ELF

Functions

create_spv

Constructs an SPV (Simplified Payment Verification) proof.

prove_bridge_circuit

Generates a Groth16 proof for the Bridge Circuit after performing sanity checks.

prove_work_only_header_chain_proof

Generates a Groth16 proof of a bitcoin header chain proof where it only outputs the total work.