bridge_circuit_host::bridge_circuit_host
Function create_spv

pub fn create_spv(
    payout_tx: Transaction,
    block_hash_bytes: &[[u8; 32]],
    payment_block: Block,
    payment_block_height: u32,
    genesis_block_height: u32,
    payment_tx_index: u32,
) -> SPV

Constructs an SPV (Simplified Payment Verification) proof.

This function decodes a Bitcoin transaction, processes block headers, constructs an MMR (Merkle Mountain Range) for block header commitment, and generates a Merkle proof for the payout transaction’s inclusion in the block.

Arguments

payout_tx - A mutable reference to a byte slice representing the payout transaction.

headers - A byte slice containing block headers, each 80 bytes long.

payment_block - A byte slice representing the full payment block.

payment_block_height - The height of the payment block in the blockchain.

payment_tx_index - The index of the payout transaction in the block’s transaction list.

Returns

Returns an SPV struct containing:

The decoded payout transaction.

A Merkle proof of the transaction’s inclusion in the block.

The block header.

An MMR proof of the block header’s inclusion in the MMR.

Panics

This function will panic if:

Decoding payout_tx or payment_block fails.

Any block header chunk is not 80 bytes long.

Generating the Merkle or MMR proof fails.