bridge_circuit_host::docker

Function pull_or_load_image

fn pull_or_load_image(
    image_digest: &str,
    container_name: &str,
    image_config_digest: &str,
) -> Result<PathBuf>

Pulls the image or loads the image from the cache if it exists. It ensures the config digest matches the given digest, or repulls the image if it doesn’t. Returns the path to the modified tar file which deleted the symlinks.