bridge_circuit_host
Function fetch_light_client_proof

pub async fn fetch_light_client_proof(
    l1_height: u32,
    client: RpcClient,
) -> Result<(LightClientProof, Receipt), ()>

Fetches the light client proof for a given L1 block height.

This function queries an Citrea light client prover RPC endpoint to obtain a light client proof for a specified L1 height. The proof is then decoded and deserialized into an InnerReceipt, which is further processed to extract a valid Receipt.

Arguments

l1_height - A u32 representing the L1 block height for which the proof is requested.

client - An instance of RpcClient used to send the JSON-RPC request.

Returns

Ok((LightClientProof, Receipt)) - A tuple containing:

LightClientProof: The extracted proof with journal bytes and L2 height.

Receipt: The transaction receipt parsed from the proof.

Panics

This function will panic if:

The RPC response does not contain a valid "proof" string.

The proof is not a valid hex string.

The proof cannot be deserialized into InnerReceipt.

The receipt cannot be extracted from the InnerReceipt.

The RPC response does not contain a valid "lastL2Height" string.

Example:

use alloy_rpc_client::RpcClient;
use alloy_rpc_client::ClientBuilder;
use bridge_circuit_host::fetch_light_client_proof;

const LIGHT_CLIENT_PROVER_URL: &str = "https://light-client-prover.testnet.citrea.xyz/";

#[tokio::main]
async fn main() {
    let light_client_rpc_client =
        ClientBuilder::default().http(LIGHT_CLIENT_PROVER_URL.parse().unwrap());

    let (light_client_proof, lcp_receipt) =
        fetch_light_client_proof(72471, light_client_rpc_client)
            .await
            .unwrap();
}