bridge_circuit_host

Function fetch_storage_proof

pub async fn fetch_storage_proof(
    l2_height: &String,
    deposit_index: u32,
    client: RpcClient,
) -> StorageProof

Fetches the storage proof for a given deposit index and transaction ID.

This function interacts with an Citrea RPC endpoint to retrieve a storage proof, which includes proof details for both the UTXO and the deposit index.

Arguments

l2_height - A reference to String representing the L2 block height. (e.g. “0x123a”)

deposit_index - A u32 representing the deposit index.

client - An instance of RpcClient used to make the JSON-RPC request.

Returns

Returns a StorageProof struct containing serialized storage proofs for the UTXO and deposit index.

Errors

This function will panic if:

keccak256(UTXOS_STORAGE_INDEX) does not return a valid 32-byte slice.

The RPC request to eth_getProof fails.

The response from the RPC call cannot be deserialized into an EIP1186AccountProofResponse.

Example:

use alloy_rpc_client::ClientBuilder;
use bridge_circuit_host::fetch_storage_proof;
use hex_literal::hex;

const CITREA_TESTNET_RPC: &str = "https://rpc.testnet.citrea.xyz/";

#[tokio::main]
async fn main() {
    let citrea_rpc_client = ClientBuilder::default().http(CITREA_TESTNET_RPC.parse().unwrap());

    let storage_proof = fetch_storage_proof(
        &"0xabc".to_string(),
        37,
        hex!("BB25103468A467382ED9F585129AD40331B54425155D6F0FAE8C799391EE2E7F"),
        citrea_rpc_client,
    )
    .await;
}