bridge_circuit_host

Module utils

Functions

calculate_succinct_output_prefix

Sha256(control_root, pre_state_digest, post_state_digest, id_bn254_fr) TODO: This function may be the same with something that is already here

get_ark_verifying_key

This is the test Verifying Key of the STARK-to-BitVM2 Groth16 proof Circom circuit.

get_ark_verifying_key_dev_mode_bridge

This is the risc0 dev mode Verifying Key of the STARK-to-BitVM2 Groth16 proof Circom circuit. The circuit doesnt verify the succint proof.

reverse_bits_and_copy