bridge_circuit_host

Module utils

Functions

calculate_succinct_output_prefix

Sha256(control_root, pre_state_digest, post_state_digest, id_bn254_fr)

digest_to_decimal 🔒

get_ark_verifying_key_dev_mode_bridge

This is the risc0 dev mode Verifying Key of the STARK-to-BitVM2 Groth16 proof Circom circuit. The circuit doesn’t verify the succinct proof.

get_ark_verifying_key_prod

This is the test Verifying Key of the STARK-to-BitVM2 Groth16 proof Circom circuit.

get_verifying_key

is_dev_mode

to_decimal 🔒

to_json

Convert a recursion VM seal (i.e. succinct receipt) into a JSON format compatible with the stark_verify witness generator. Taken from risc0-groth16 v2.3.2. This function will be removed once Risc Zero publishes a new version that exposes this function publicly.

total_work_from_wt_tx